refactor: divide Profile.jsx into small components for readibility and scalability
