+++ frontend/src/pages/Profile/ProfileSavedPost.jsx
@@ -1,0 +1,34 @@
+import { useNavigate } from "react-router-dom";
+
+function ProfileSavedPost({user}) {
+  // Post
+  const navigate = useNavigate()
+  const handlePostClick = (postId) => {
+    navigate(`/forum/${postId}`);
+  };
+
+  return (
+    <div className='mainS'>
+      <div className="posts-container" style={{ backgroundColor: "#333333" }}>
+        {user.savedPosts.slice().reverse().map((post, index) => {
+          const date = new Date(post.createdAt);
+          const formattedDate = `${date.getDate()} ${date.toLocaleString('en', { month: 'long' })}, ${date.getFullYear()}`;
+          return (
+            <div className='spost'>
+              <div className='post-card' key={index} style={{ backgroundColor: "#000000" }}>
+                <div className='post-icon-title'>
+                  <img src="/src/assets/save.svg" alt="" width={30} height={30} />
+                  <div className='post-title' key={post._id} onClick={() => handlePostClick(post._id)}>{post.title}</div>
+                </div>
+
+                <div className='post-date'>{formattedDate}</div>
+              </div>
+            </div>
+          );
+        })}
+      </div>
+    </div>
+  )
+}
+
+export default ProfileSavedPost
